revert: use split
